- tucker3: minor changes typedefs of coefficients
git-svn-id: https://vmmlib.svn.sourceforge.net/svnroot/vmmlib/trunk@373 4d770e62-bb1e-0410-ad62-f2382d3e8790

+++ tests/tucker3_tensor_test.cpp
@@ -249,7 +249,8 @@ namespace vmml
 		
 		tucker3_tensor< 2, 2, 2, 3, 2, 2, unsigned char, unsigned short > tuck3_hooi_3( core_hooi_3, u1_hooi_3, u2_hooi_3, u3_hooi_3 );
 		
-		double u1_min, u1_max, u2_min, u2_max, u3_min, u3_max, core_min, core_max;
+		
+		float u1_min, u1_max, u2_min, u2_max, u3_min, u3_max, core_min, core_max;
 		
 		tuck3_hooi_3.enable_quantify_coeff();
 		tuck3_hooi_3.decompose( t3_data_hooi_3, u1_min, u1_max, u2_min, u2_max, u3_min, u3_max, core_min, core_max );
